python-api/endpoint.py
from flask import Flask, request, jsonify

app = Flask(__name__)

# Endpoint to create a new event
@app.route('/events', methods=['POST'])
def create_event():
    # Get the request data
    data = request.get_json()

    # Extract the event parameters
    event_date = data['event_date']
    location = data['location']
    title = data['title']
    description = data['description']
    category = data['category']

    # TODO: Insert the event data into the database

    # Prepare the response
    response = {'message': 'Event created successfully'}

    # Return the response as JSON
    return jsonify(response)

# Endpoint to get all events
@app.route('/events', methods=['GET'])
def get_all_events():
    # TODO: Retrieve all events from the database

    # Prepare the response
    response = {'message': 'All events retrieved successfully', 'events': []}

    # TODO: Add retrieved events to the response

    # Return the response as JSON
    return jsonify(response)

# Endpoint to get events by category
@app.route('/events/category/<category>', methods=['GET'])
def get_events_by_category(category):
    # TODO: Retrieve events from the database based on the category

    # Prepare the response
    response = {'message': f'Events under category {category} retrieved successfully', 'events': []}

    # TODO: Add retrieved events to the response

    # Return the response as JSON
    return jsonify(response)

# Endpoint to get events by location
@app.route('/events/location/<location>', methods=['GET'])
def get_events_by_location(location):
    # TODO: Retrieve events from the database based on the location

    # Prepare the response
    response = {'message': f'Events in location {location} retrieved successfully', 'events': []}

    # TODO: Add retrieved events to the response

    # Return the response as JSON
    return jsonify(response)

# Endpoint to get events by filter (any field)
@app.route('/events/filter', methods=['GET'])
def get_events_by_filter():
    # Get the filter parameter from the query string
    filter_param = request.args.get('filter')

    # TODO: Retrieve events from the database based on the filter

    # Prepare the response
    response = {'message': f'Events filtered by {filter_param} retrieved successfully', 'events': []}

    # TODO: Add retrieved events to the response

    # Return the response as JSON
    return jsonify(response)

# Endpoint to delete an event
@app.route('/events/<event_id>', methods=['DELETE'])
def delete_event(event_id):
    # TODO: Delete the event from the database based on the event ID

    # Prepare the response
    response = {'message': f'Event with ID {event_id} deleted successfully'}

    # Return the response as JSON
    return jsonify(response)

if __name__ == '__main__':
    app.run()
